+++ examples/tokenvm/rpc/dependencies.go
@@ -1,0 +1,4 @@
+package rpc
+
+type Controller interface {
+}
